refactor: memberSlice를 authSlice로 명칭을 변경한다
