+++ front-end/src/domains/UserNavigation/UserNavigation.tsx
@@ -10,7 +10,7 @@ import {
 } from "./UserNavigation.styled";
 import { PATH } from "../../constants";
 import { useAppDispatch } from "../../app/hooks";
-import { logout } from "../../features/member/memberSlice";
+import { logout } from "../../features/member/authSlice";
 
 type NavigationProps = HTMLAttributes<HTMLElement>;
