+++ front-end/src/routes/PrivateRoute/PrivateRoute.tsx
@@ -2,7 +2,7 @@ import { ComponentProps } from "react";
 import { Redirect, Route } from "react-router-dom";
 import { PATH } from "../../constants";
 import { useAppSelector } from "../../app/hooks";
-import { selectIsAuthenticated } from "../../features/member/memberSlice";
+import { selectIsAuthenticated } from "../../features/member/authSlice";
 
 const PrivateRoute = ({ children, ...rest }: ComponentProps<typeof Route>) => {
   const isAuthenticated = useAppSelector(selectIsAuthenticated);
